* 7.0:
  fix some compiler warnings
  [eEPGCache] use PyExc_TypeError with readable text and remove eDebug
  [eDVBCIInterfaces] use PyExc_ValueError
  [eDVBDB] use PyExc_TypeError with readable text and remove eDebug
  [eDVBResourceManager] use PyExc_TypeError
  fix compiler warning:
  remove c++ python2 support
  remove PyEval_InitThreads() for py3
  translation update
  Improve and optimise SystemInfo code (#2320)
  [skin.py] Remove superfluous reloadWindoStyles() call
  Hotkey record event (#2313) (#2317)
  [SystemInfo.py] Correct typo (#2314)

+++ lib/service/servicem2ts.cpp
@@ -277,7 +277,7 @@ sync:
 		if (tmp[4] != 0x47)
 		{
 			if (rd > 0) {
-				eDebug("[eM2TSFile] short read at pos %lld async!!", m_current_offset);
+				eDebug("[eM2TSFile] short read at pos %jd async!!", (intmax_t)m_current_offset);
 				return rd;
 			}
 			else {
